Android: Don't kill app when SIM card or density changes
This adds subscription to additional config changes on Android to
avoid the application automatically restarting when the SIM card is
replaced or the logical density of the screen is changed (as a result
of the user changing it in settings).

These restarts are especially problematic when using Qt in a
home screen application due to the way they are launched.

[ChangeLog][Android] Avoid the system restarting application when
user replaces SIM card or changes the logical density of the screen.

Task-number: QTBUG-63735
Change-Id: Iba85fe6239b2d29b543b3c4e2f953efb9b54f713
Reviewed-by: BogDan Vatra <bogdan@kdab.com>