QHeaderView: fix inconsistent saved state, ignored during restore
This happens because QTreeView disconnects QHeaderView's
_q_layoutChanged slot (!). So the stretching of the last section,
done in _q_layoutAboutToBeChanged, invalidated total length, but didn't
recalculate it (since that's done in _q_layoutChanged). As a result,
length was inconsistent, and saveState would save that, and
restoreState() would early-return, not doing anything.

Let's just ensure length is always consistent, so we don't depend on the
complex issue of whether _q_layoutChanged should be called or not.

This an adapted backport of 4a04eea4f4 from 5.11, the unittest shows that
c0e45ae851 is missing in this branch though.

Change-Id: I4137a19e0a6fdf820dd53fb55e858d1d04a2c113
Reviewed-by: Thorbjørn Lund Martsum <tmartsum@gmail.com>